# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register new user",
        "description": "",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserRegister"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created: You created an account"
          },
          "400": {
            "description": "Bad Request: Invalid or missing credentials"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while registering the user"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User login",
        "description": "NOTE: After logging in as admin you get a cookie token and will be able to use the other endpoints",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK: You are logged in"
          },
          "401": {
            "description": "Bad Request: Incorrect email or password"
          },
          "500": {
            "description": "Internal Server Error: Something went wrong with creating JWT token"
          }
        }
      }
    },
...
    "/admin/brands": {
      "get": {
        "tags": [
          "Admin Panel"
        ],
        "summary": "Fetches data for brands page",
        "description": "You need to execute /auth/login as admin first",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admins have access"
          },
          "500": {
            "description": "Internal Server Error: Error fetching data for brands page"
          }
        }
      }
    },
...
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "summary": "View all brands",
        "description": "",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to view brands"
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "summary": "Adds a new brand.",
        "description": "You need to execute /auth/login as admin first",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PostBrand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created: Brand created successfully"
          },
          "400": {
            "description": "Bad Request:Invalid or missing properties"
          },
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admins have access"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to add brand"
          }
        }
      }
    },
    "/brands/{id}": {
      "put": {
        "tags": [
          "Brands"
        ],
        "summary": "Updates a brand by ID",
        "description": "You need to execute /auth/login as admin first",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateBrand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created: Brand updated successfully"
          },
          "400": {
            "description": "Bad Request: Invalid properties"
          },
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admins have access"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to update brand"
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "summary": "Deletes a brand by ID",
        "description": "You need to execute /auth/login as admin first",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK: Brand deleted successfully"
          },
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admins have access"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict: You are trying to delete a brand thats in use"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to delete brand"
          }
        }
      }
    },
...
  },
  "definitions": {
    "UserRegister": {
      "type": "object",
      "properties": {
        "firstname": {
          "type": "string",
          "example": "John"
        },
        "lastname": {
          "type": "string",
          "example": "Doe"
        },
        "username": {
          "type": "string",
          "example": "JohnDoe90"
        },
        "email": {
          "type": "string",
          "example": "john.doe@example.com"
        },
        "password": {
          "type": "string",
          "example": "abcd1234"
        },
        "address": {
          "type": "string",
          "example": "Main Street 10"
        },
        "telephone": {
          "type": "string",
          "example": "12345678"
        }
      }
    },
    "UserLogin": {
      "type": "object",
      "properties": {
        "emailOrUsername": {
          "type": "string",
          "example": "admin"
        },
        "password": {
          "type": "string",
          "example": "P@ssword2023"
        }
      }
    },
...
    "PostBrand": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "New brand"
        }
      }
    },
...
  }
}